requirements.txt
pymongo
flask-restful
tqdm
requests
inflect
dotmap==1.3.8
PyYAML==5.1.1
pandas
google-api-python-client
nltk